requirements.txt
# Core API
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Frontend
streamlit==1.28.1

# AI/ML
google-generativeai==0.3.2
scikit-learn==1.3.2
numpy==1.24.3

# Document Processing
PyPDF2==3.0.1
python-docx==0.8.11
beautifulsoup4==4.12.2
markdown==3.5.1
pdfplumber==0.9.0
PyMuPDF==1.23.8
pdfminer.six==20221105

# OCR (optional)
pytesseract==0.3.10
Pillow==10.1.0

# Text Processing
nltk==3.8.1

# Utilities
requests==2.31.0

# Visualization and Analysis
plotly==5.17.0
pandas==2.1.3
psutil==5.9.6
